generate_instructions/requirements.txt
clip
lxml
matplotlib
numpy
tqdm
torch==1.7.1
torchvision==0.8.2
pycocotools
Pillow
